Add iNES file display. (#11)
* Split PCXEditorCanvas into the editor and the visualizer.

* Finish new nes file display.

* Add scroll to the header section of the NESFileEditor.

* Rename name tables' section.

* Improve memory table.

* Now nametables display palette colors.

* NES4JAMS now supports the latest JAMS version.

* Change sprites to assemble spanish name.

+++ src/main/kotlin/io/github/gaeqs/nes4jams/gui/simulation/memory/representation/NESNumberRepresentation.kt
@@ -69,7 +69,7 @@ abstract class NESNumberRepresentation(
     override fun getName() = name
     override fun getResourceProvider(): ResourceProvider = NES4JAMS.INSTANCE
 
-    fun represent(address: UShort, pane: NESMemoryPane): String {
+    fun represent(address: UInt, pane: NESMemoryPane): String {
         var first: UByte = 0u
         var second: UByte = 0u
         pane.simulation.runSynchronized {
